requirements.txt
torch>=2.0
numpy>=1.20
scipy>=1.7
Pillow>=9.0
